rename installer
