+++ tipo_installer.py
